+++ CHANGELOG.md
@@ -40,6 +40,11 @@
     - This component contains a single bool which denotes whether you have authority over that component.
     - It will not tell you about soft-handover (`AuthorityLossImminent`).
 - You may now `[Require]` `EntityId`, `Entity`, `World`, `ILogDispatcher`, and `WorldCommandSender` in Monobehaviours.
+- Added constructors for all generated component snapshot types.
+- Added the ability to send arbitrary serialized data in a player creation request.
+    - Replaced `Vector3f` position in `CreatePlayerRequestType` with a `bytes` field for sending arbitrary serialized data.
+    - Added `SerializeArguments` and `DeserializeArguments<T>` methods to `PlayerLifecycleHelper`.
+- Added `RequestPlayerCreation` to manually request for player creation in `SendCreatePlayerRequestSystem`.
 - Added a menu item, navigate to **SpatialOS** > **Generate Dev Authentication Token**, to generate a TextAsset containing the [Development Authentication Token](https://docs.improbable.io/reference/latest/shared/auth/development-authentication).
 
 ### Changed
@@ -54,6 +59,7 @@
 
 - Fixed a bug where running `SpatialOS -> Generate code` would always regenerate code, even if no files had changed.
 - Fixed a bug where building all workers in our sample projects would fail, if you have Android build support installed, but didn't set the path to the Android SDK.
+- Fixed a bug where some prefabs would not be processed correctly, causing NullReferenceExceptions in OnEnable.
 
 ### Internal